% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_var.R
\name{find_var}
\alias{find_var}
\title{Find variables}
\usage{
find_var(x, tables = c("students", "dates", "romances", "hookups"))
}
\arguments{
\item{x}{A pattern to search (grep) for}

\item{tables}{A character vector of table names to search}
}
\description{
Search for a pattern in the OCSLS table column names
}
